Move madx optics file generation to module
