+++ xtrack/_temp/lhc_match/__init__.py
@@ -1,2 +1,3 @@
 from .lhc_match import *
 from .var_limits import set_var_limits_and_steps
+from .gen_madx_optics_file import gen_madx_optics_file_auto
